*refactoring of ConvertMaxPoolNode.
Move ConvertMaxPoolNode from OnnxRuntime.h into ConvertMaxPoolNode.cpp
and register the new source in the CvtOnnx Visual Studio project.
